functions/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
functions-framework>=3.5.0
google-cloud-firestore>=2.16.0
firebase-admin>=6.5.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
httpx>=0.27.0
python-dotenv>=1.0.1
polyline>=2.0.2
mangum>=0.17.0
